chore: move sendStoredEmail
